orientation/requirements.txt
Flask==0.12.2
flask-cors==3.0.3
pyserial==3.4
waitress==1.1.0
